Move topology provider and requests to own crate
