+++ Cargo.toml
@@ -74,6 +74,7 @@ members = [
     "common/types",
     "common/wasm-utils",
     "explorer-api",
+    "explorer-api/explorer-api-requests",
     "gateway",
     "gateway/gateway-requests",
     "integrations/bity",
@@ -117,31 +118,31 @@ anyhow = "1.0.71"
 async-trait = "0.1.64"
 bip39 = { version = "2.0.0", features = ["zeroize"] }
 cfg-if = "1.0.0"
+cosmrs = "=0.14.0"
 cosmwasm-derive = "=1.2.5"
 cosmwasm-schema = "=1.2.5"
 cosmwasm-std = "=1.2.5"
 cosmwasm-storage = "=1.2.5"
-cosmrs = "=0.14.0"
-# same version as used by cosmrs
-tendermint-rpc = "=0.32"
-cw-utils = "=1.0.1"
+cw-controllers = { version = "=1.0.1" }
 cw-storage-plus = "=1.0.1"
+cw-utils = "=1.0.1"
 cw2 = { version = "=1.0.1" }
 cw3 = { version = "=1.0.1" }
 cw3-fixed-multisig = { version = "=1.0.1" }
 cw4 = { version = "=1.0.1" }
-cw-controllers = { version = "=1.0.1" }
 dotenvy = "0.15.6"
 generic-array = "0.14.7"
-k256 = "0.13"
 getrandom = "0.2.10"
+k256 = "0.13"
 lazy_static = "1.4.0"
 log = "0.4"
 once_cell = "1.7.2"
 rand = "0.8.5"
+reqwest = "0.11.18"
 serde = "1.0.152"
 serde_json = "1.0.91"
 tap = "1.0.1"
+tendermint-rpc = "=0.32" # same version as used by cosmrs
 thiserror = "1.0.38"
 tokio = "1.24.1"
 url = "2.2"
