+++ common/client-core/Cargo.toml
@@ -10,27 +10,29 @@ rust-version = "1.66"
 [dependencies]
 async-trait = { workspace = true }
 base64 = "0.21.2"
-dirs = "4.0"
 dashmap = "5.4.0"
+dirs = "4.0"
 futures = "0.3"
 humantime-serde = "1.0"
 log = { workspace = true }
 rand = { version = "0.7.3", features = ["wasm-bindgen"] }
+reqwest = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 sha2 = "0.10.6"
 tap = "1.0.1"
 thiserror = "1.0.34"
-url = { version ="2.2", features = ["serde"] }
-tungstenite = { version = "0.13.0", default-features = false }
-tokio = { version = "1.24.1", features = ["macros"]}
 time = "0.3.17"
+tokio = { version = "1.24.1", features = ["macros"]}
+tungstenite = { version = "0.13.0", default-features = false }
+url = { version ="2.2", features = ["serde"] }
 zeroize = { workspace = true }
 
 # internal
 nym-bandwidth-controller = { path = "../bandwidth-controller" }
 nym-config = { path = "../config" }
 nym-crypto = { path = "../crypto" }
+nym-explorer-api-requests = { path = "../../explorer-api/explorer-api-requests" }
 nym-gateway-client = { path = "../client-libs/gateway-client" }
 #gateway-client = { path = "../../common/client-libs/gateway-client", default-features = false, features = ["wasm", "coconut"] }
 nym-gateway-requests = { path = "../../gateway/gateway-requests" }
