+++ explorer-api/explorer-api-requests/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "nym-explorer-api-requests"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+nym-contracts-common = { path = "../../common/cosmwasm-smart-contracts/contracts-common" }
+nym-mixnet-contract-common = { path = "../../common/cosmwasm-smart-contracts/mixnet-contract" }
+nym-validator-client = { path = "../../common/client-libs/validator-client" }
+schemars = { version = "0.8", features = ["preserve_order"] }
+serde = { version = "1.0", features = ["derive"] }
+ts-rs = { version = "6.1.2", optional = true }
